chore: Refactoring jsonschema imports/exports
