+++ packages/common/src/jsonschema/constants/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Metadata key
+ * @private
+ * @type {string}
+ */
+export const PROPERTIES_METADATA = "tsed:properties:metadata";
