+++ packages/common/src/jsonschema/index.ts
@@ -1,6 +1,7 @@
 export * from "./decorators";
 export * from "./class/JsonSchema";
 export * from "./class/PropertyMetadata";
+export * from "./interfaces/IPropertyOptions";
 export * from "./registries/PropertyRegistry";
 export * from "./registries/JsonSchemesRegistry";
 export * from "./services/JsonSchemesService";
