+++ packages/common/src/jsonschema/interfaces/IPropertyOptions.ts
@@ -1,0 +1,6 @@
+import {Type} from "@tsed/core";
+
+export interface IPropertyOptions {
+  name?: string;
+  use?: Type<any>;
+}
